requirements.txt
Flask
gunicorn
numpy
tensorflow
Pillow
qrcode